TE: Refactor order's scripts
