+++ tests/E2E/test/campaigns/high/01_orders/2_create_order_in_FO.js
@@ -1,0 +1,70 @@
+const {AccessPageFO} = require('../../../selectors/FO/access_page');
+const {productPage}= require('../../../selectors/FO/product_page');
+const {CheckoutOrderPage}= require('../../../selectors/FO/order_page');
+const {OrderPage} = require('../../../selectors/BO/order');
+const {AccessPageBO} = require('../../../selectors/BO/access_page');
+let promise = Promise.resolve();
+
+scenario('Create order in FO', () => {
+
+    scenario('Open the browser and connect to the FO', client => {
+        test('should open the browser', () => client.open());
+        test('should sign in FO', () => client.signInFO(AccessPageFO));
+    }, 'order');
+
+    scenario('Create order in FO', client => {
+        test('should change the FO language to english', () => client.changeLanguage());
+        test('should choose product ', () => client.waitForExistAndClick(productPage.first_product));
+        test('should select product "size M" ', () => client.waitAndSelectByValue(productPage.first_product_size, '2'));
+        test('should select product "color blue"', () => client.waitForExistAndClick(productPage.first_product_color));
+        test('should set the product "quantity"', () => client.waitAndSetValue(productPage.first_product_quantity, "4"));
+        test('should click on "Add to cart" button', () => client.waitForExistAndClick(CheckoutOrderPage.add_to_cart_button));
+        test('should click on "Proceed to checkout" button from "Product" page', () => client.waitForExistAndClick(CheckoutOrderPage.proceed_to_checkout_modal_button));
+        test('should click on "Proceed to checkout" button from "Cart" page', () => client.waitForExistAndClick(CheckoutOrderPage.proceed_to_checkout_button));
+        test('should click on confirm address button', () => client.waitForExistAndClick(CheckoutOrderPage.checkout_step2_continue_button));
+        test('should choose shipping method my carrier', () => client.waitForExistAndClick(CheckoutOrderPage.shipping_method_option));
+        test('should create message', () => client.waitAndSetValue(CheckoutOrderPage.message_textarea, 'Order message test'));
+        test('should click on "confirm delivery" button', () => client.waitForExistAndClick(CheckoutOrderPage.checkout_step3_continue_button));
+        test('should set the payment type "Payment by bank wire"', () => client.waitForExistAndClick(CheckoutOrderPage.checkout_step4_payment_radio));
+        test('should set "the condition to approve"', () => client.waitForExistAndClick(CheckoutOrderPage.condition_check_box));
+        test('should click on order with an obligation to pay button', () => client.waitForExistAndClick(CheckoutOrderPage.confirmation_order_button));
+        test('should check the order confirmation', () => {
+            return promise
+                .then(() => client.checkTextValue(CheckoutOrderPage.confirmation_order_message, 'YOUR ORDER IS CONFIRMED', "contain"))
+                .then(() => client.getTextInVar(CheckoutOrderPage.order_product, "product"))
+                .then(() => client.getTextInVar(CheckoutOrderPage.order_basic_price, "basic_price"))
+                .then(() => client.getTextInVar(CheckoutOrderPage.order_total_price, "total_price"))
+                .then(() => client.getTextInVar(CheckoutOrderPage.order_reference, "reference", true))
+                .then(() => client.getTextInVar(CheckoutOrderPage.shipping_method, "method", true))
+                .then(() => client.getTextInVar(CheckoutOrderPage.order_shipping_prince_value, "shipping_price"))
+        });
+        test('should logout successfully from the Front Office', () => client.signOutFO(AccessPageFO));
+    }, 'order', true);
+
+  scenario('Check the created order in BO', () => {
+    scenario('Open the browser and connect to the BO', client => {
+      test('should open the browser', () => client.open());
+      test('should log in successfully in Back Office', () => client.signInBO(AccessPageBO));
+    }, 'order');
+
+    scenario('Check the created order in BO', client => {
+      test('should go to "Orders" page', () => client.goToSubtabMenuPage(OrderPage.orders_subtab, OrderPage.order_submenu));
+      test('should search the order created by reference', () => client.waitAndSetValue(OrderPage.search_by_reference_input, global.tab['reference']));
+      test('should go to search order', () => client.waitForExistAndClick(OrderPage.search_order_button));
+      test('should go to the order ', () => client.scrollWaitForExistAndClick(OrderPage.view_order_button.replace('%NUMBER', 1)));
+      test('should check the customer name ', () => client.checkTextValue(OrderPage.customer_name, 'John DOE', 'contain'));
+      test('should status be equal to Awaiting bank wire payment ', () => client.checkTextValue(OrderPage.order_status, 'Awaiting bank wire payment'));
+      test('should check the shipping price', () => client.checkTextValue(OrderPage.shipping_cost, global.tab['shipping_price']));
+      test('should check the product', () => client.checkTextValue(OrderPage.product_name, global.tab['product']));
+      test('should check the order message ', () => client.checkTextValue(OrderPage.message_order, 'Order message test'));
+      test('should check the total price', () => client.checkTextValue(OrderPage.total_price, global.tab["total_price"]));
+      test('should check basic product price', () => {
+        return promise
+          .then(() => client.scrollTo(OrderPage.edit_product_button))
+          .then(() => client.waitForExistAndClick(OrderPage.edit_product_button))
+          .then(() => client.checkAttributeValue(OrderPage.product_basic_price, 'value', global.tab["basic_price"].replace('€', '')))
+      });
+      test('should check shipping method ', () => client.checkTextValue(OrderPage.shipping_method, global.tab["method"].split('\n')[0], 'contain'));
+    }, "order");
+  }, "order");
+}, 'order', true);
