+++ tests/E2E/test/campaigns/high/02_product/6_pack_stock.js
@@ -3,9 +3,8 @@ const {AccessPageBO} = require('../../../selectors/BO/access_page');
 const {ShopParametre} = require('../../../selectors/BO/shopParameters/');
 const {ProductSettings} = require('../../../selectors/BO/shopParameters/product_settings');
 const {productPage}= require('../../../selectors/FO/product_page');
-const {buyOrderPage}= require('../../../selectors/FO/buy_order_page');
+const {CheckoutOrderPage}= require('../../../selectors/FO/order_page');
 const {SearchProductPage} = require('../../../selectors/FO/search_product_page');
-const {layerCart}= require('../../../selectors/FO/layer_cart_page');
 const {AddProductPage} = require('../../../selectors/BO/add_product_page');
 const common_scenarios = require('../02_product/product');
 
@@ -51,13 +50,12 @@ scenario('Check "Orders"', client => {
     test('should search for the product "A"', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, productData[0]['name'] + date_time));
     test('should go to the product page', () => client.waitForExistAndClick(SearchProductPage.product_result_name));
     test('should set the product "quantity"', () => client.waitAndSetValue(productPage.first_product_quantity, "50"));
-    test('should click on "Add to cart" button  ', () => client.waitForExistAndClick(buyOrderPage.add_to_cart_button));
-    test('should verify the appearance of the green confirmation', () => client.checkGreenConfirmation());
-    test('should click on proceed to checkout button', () => client.waitForExistAndClick(layerCart.command_button));
-  }, 'order/order');
+    test('should click on "Add to cart" button  ', () => client.waitForExistAndClick(CheckoutOrderPage.add_to_cart_button));
+    test('should click on proceed to checkout button', () => client.waitForVisibleAndClick(CheckoutOrderPage.proceed_to_checkout_modal_button));
+  }, 'order');
   scenario('Create order with 10 item of pack B ', client => {
     test('should search for the product', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, productData[1]['name'] + date_time));
     test('should go to the product page', () => client.waitForExistAndClick(SearchProductPage.product_result_name));
-    test('should verify that the button "AJOUTER AU PANIER" is DISABLED', () => client.checkEnable(buyOrderPage.add_to_cart_button));
-  }, 'order/order');
+    test('should verify that the button "AJOUTER AU PANIER" is DISABLED', () => client.checkEnable(CheckoutOrderPage.add_to_cart_button));
+  }, 'order');
 }, 'product/product',true);
