+++ tests/E2E/test/selectors/BO/index.js
@@ -5,8 +5,7 @@ module.exports = Object.assign({},
   require('./add_product_page'),
   require('./module_page'),
   require('./onboarding'),
-  require('./order_page'),
-  require('./create_order'),
+  require('./order'),
   require('./product_list'),
   require('./shopParameters'),
   require('./international')
