--- tests/E2E/test/selectors/FO/buy_order_page.js
@@ -1,37 +1,0 @@
-module.exports = {
-  buyOrderPage: {
-    add_to_cart_button: '//*[@id="add-to-cart-or-refresh"]/div[2]/div/div[2]/button',
-    green_confirmation: '//*[@id="myModalLabel"]',
-    first_product_home_page_name: '[itemprop="name"]',
-    product_image: '#content',
-    product_name_details: '[itemprop="name"]',
-    product_price_details: '[itemprop="price"]',
-    product_quantity_details: '#quantity_wanted',
-    product_name: '(//div[@class="product-line-info"])[1]/a',
-    product_price: '//span[@class="price"]',
-    checkout_step2_continue_button: '//*[@id="checkout-addresses-step"]/div/div/form/div[2]/button',
-    checkout_step3_continue_button: '//*[@id="js-delivery"]/button',
-    checkout_step4_payment_radio: '//*[@id="payment-option-2"]',
-    checkout_step4_cgv_checkbox: '//input[@id="conditions_to_approve[terms-and-conditions]"]',
-    checkout_step4_order_button: '#payment-confirmation >div > button',
-    checkout_total: '//div[@class="cart-summary-line cart-total"]/span[2]',
-    order_confirmation_name: '#order-items > div > div > div.col-sm-4.col-xs-9.details > span',
-    order_confirmation_price1: '#order-items > div > table > tbody > tr:nth-child(1) > td:nth-child(2)',
-    order_confirmation_price2: '#content-hook_payment_return > div > div > div > dl > dd:nth-child(2)',
-    order_confirmation_ref: '(//div[@id="order-details"]/ul/li)[1]',
-    shipping_method_option: '//*[@id="delivery_option_2"]',
-    message_textarea: '//*[@id="delivery_message"]',
-    condition_check_box: '//*[@id="conditions_to_approve[terms-and-conditions]"]',
-    confirmation_order_button: '//*[@id="payment-confirmation"]/div[1]/button',
-    confirmation_order_message: '//*[@id="content-hook_order_confirmation"]/div/div/div/h3',
-    order_product: '//*[@id="order-items"]/div/div/div[2]/span',
-    order_reference: '//*[@id="order-details"]/ul/li[1]',
-    order_basic_price: '//*[@id="order-items"]/div/div/div[3]/div/div[1]',
-    order_total_price: '//*[@id="order-items"]/div/table/tbody/tr[1]/td[2]',
-    order_shipping_prince_value: '//*[@id="order-items"]/div/table/tbody/tr[2]/td[2]',
-    customer_name: '//*[@id="_desktop_user_info"]/div/a[2]/span',
-    shipping_method: '//*[@id="order-details"]/ul/li[3]',
-
-
-  }
-};
