--- tests/E2E/test/selectors/FO/layer_cart_page.js
@@ -1,12 +1,0 @@
-module.exports = {
-  layerCart:{
-  layer_cart: '//div[@id="blockcart-modal" and @style="display: block;"]',
-  name_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/h6',
-  price_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/p[1]',
-  quantity_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/p[2]',
-  command_button: '//*[@id="blockcart-modal"]/div/div/div[2]/div/div[2]/div/div/a',
-  proceed_to_checkout_button: '//*[@id="main"]/div/div[2]/div[1]/div[2]/div/a',
-
-
-  }
-};
